rebased dynamic apploader
Added dynamic app loader to kernel and capsule. Created two components to enable initialization of capsule and kernel logic. Made changes to load_processes() and load_processes_from_flash() functions to return remaining memory.

fixed warnings

fixed some warnings causing CI fail

fixed some format issues

clippy fixes

fixed header check issue

resolved some pr comments

pr comments changes

changed it so that the capsule sends a subslice of the buffer to the kernel so the kernel does not have to compare lengths that the capsule provided vs the length of the buffer the capsule actually sent. Generally a safer approach.

Improved State Machine

Improved the state machine to better track if a userland app is requesting write to the bounds allocated to it during the setup phase. This state machine also ensures that when the process loader is busy, another request cannot be made.

Improved State Machine

Improved the state machine to better track if a userland app is requesting write to the bounds allocated to it during the setup phase. Also tracks if the dynamic process loader is currently busy. Performed some code cleanup.

Header validation + Erasing app upon failure

The kernel now validates the header before writing the app into flash. In addition, upon failure at any stage, the flash region will be reclaimed for future use. Added a Fail state to help track failure modes better.

bug fixes

fixed two instances in app write when resets were not taking place. also fixed an improper condition check for declared length vs length in header.

states bug fix

An app that was successfully could not be made into a process object because of errant state tracking. That was fixed. Additionally, removed unnecessary state and parameter clears during Busy state.

fixed alignment and padding

Apps were previously aligned but we were not writing padding before new apps to match the linkedlist with the alignment. This is now fixed. A new enum was added to track the padding requirement. Additionally, the setup phase is now asynchronous to accomodate cases where you may or may not have to write padding after an app depending on if there is an app stored beyond our newly written app. This change propogates to the capsule as well.

remove extraneous function call

check_padding_requirements was called twice once during postpad and during prepad to return the next_app_start_addr and previous_app_end_addr respectively. Changed the function so that the values are stored and the function does not have to be called twice.

alignment improvement

changed the way the app address is identified during setup phase to align the app to the power of 2 based on its size and start address.